Start changing tool inputs to accept descriptions
